implement sysdep replacement as proposed by dalias
the .mo format was designed in a way such that all application
can used a single shared non-mutable copy of the translation strings
in memory.
years later an issue was identified with format string specifier macros
that differ between platforms (like PRIu64) and a new
syntax introduced: %<PRIuMAX>.

this however breaks the non-mutable aspect of the .mo files, since now
every application has to remap a modified copy of the strings containing
these "sysdep" format specifiers.
we simply replace these with all possible expanded values of such a macro
directly in the generated .mo file.