Revert "msgfmt: sort strings"
This reverts commit ee82b152ce2b4ab74b7ac72d27540f0dd01172e8.

While this now creates valid translation files, it opened the door
for a lot of nasty cornercases like plural forms that were elegantly
circumvented in the earlier design, but now causing breakage.
such breakage was experienced with de.pot from gnu coreutils.

development of the "full-featured" msgfmt will eventually continue
in the "full" branch and merged into master when it's ready for
prime-time.