[1/3]: more flexible way of installing share files following https://github.com/sabotage-linux/gettext-tiny/issues/12
this is the first part of the solution the issue, to prepare for
introducing Makefile.in.in in the next commit.

we modify m4dir to datadir, as we will introduce Makefile.in.in, which
should be copied to $datadir$/gettety-tiny. It's a better name. Also
moving m4 to a subdir makes a cleaner structure and a good place to
place Makefile.in.in.