[2/3]: introduce GNU makefile following https://github.com/sabotage-linux/gettext-tiny/issues/12
this for supporting auto generation of po Makefile. Softwares may not
write their own Makefile.in for po/* at all. Instead, they depend on
autotools to generate Makefile. And the template makefile.in.in comes
from autopoint.

But the gnu one does not suite gettext-tiny, since xgettext and msgmerge
both are stubs. So, this commit did not work, and it may cause error,
DO NOT USE! A suitable version is to be added in the next commit.