fix: writestr() should do nothing when empty .po https://github.com/sabotage-linux/gettext-tiny/issues/14.
At line:439, writestr() is called to write down those strings that
have not been wrote down in the main loop.

When po files are empty, nothing is wrote down, so d->plurals_count ==
0. Then, deleting 'invalid' codes in writestr() was executed wrongly.
-1 is not a valid index, qsort will segfault, of course.

To fix the bug, we check d->num[pe_msgid]>0, so it will do nothing when
the files are empty.