poparser: ignore the invalid after comments when compiling gdk-pixbuf 2.36.11, msgfmt segfaulted because poparser went though a situation where: ``` msgstr[1] "" "ssss "vvvv ```
In case, it extracted the string ' "ssss" ' and see the remaining part ' vvvv '
as the invalid, there's no wrong with the logic itself. But i wonder if we
could just be generous with these merged strings generated by xgettext. If we
insist on correctness everytime po files are updated by xgettext, we may meet
numbers of abort(). Seeing gnu one ignoring these obvious invalid lines, i made
this commit to provide a more smooth user experience.