xgettext: create file with empty outputdir arg Found when updating neovim.
The original one will try `touch ./...file` when outputdir arg is empty.
But it failed because cmake passed an absolute path to our script. After
my modification, it will still try to touch file in a relative path with a
empty outputdir arg, but wont fail when the outputfile arg is an
absolute path.