poparser: not to skip sysdeps other than %<PRIu32> follow https://github.com/sabotage-linux/gettext-tiny/issues/39#issuecomment-445036044.
it's obvious that, strstr will search for `%<PRIu32>` first, if there's
one, then we get there and skip all other sysdep strings before the first
`%<PRIu32>`. but what we want is, to search the first sysdep string.

so, our new stragegy is to search for `%` instead. such that, we will
always match the first sysdep string.