poparser: support feeds anything[1/2] this is the first patch. It mainly solves the case like: `msgid "ss" "gg"`.
This is a valid case(proved by GNU msgfmt). Just like C compiler, we
should concat strings, treat that string as `"ssgg"`.

Previous feed_line assume there's only one `"xxx"` string in one line.
For example, when it parses `msgid "ss"`, whole line is marked as
parsed, so `"gg"` is ignored.

Now it turns into a loop and will consume exactly the part it parsed, so
that all strings can be consumed.