poparser: support feeds anything[2/2] this is the second patch. It mainly changes the working method of feed_line().
Previously, feed_line() assumes the input is a complete whole line. Also
it uses the buffer passed by init() to do iconv things.

Now, feed_line() assumes nothing about the input. It accumulates strings
until a complete line is feeded. A postive return value says the
poparser is expecting more strings and it's not an error.

But after when you feed all the data, poparser should consumes all
tokens exactly, so the last call to feed_line() should return 0.

And that introduce the 2nd issue. Incomplete strings will not be consumed
, so it needs to be stored. Thus, we alloc an internal buffer. When we
do not need to do iconv, we just copy new strings after old strings. Or
let iconv do the copy. This also elimates the need of passing an iconv
buffer to poparser.

now msgfmt and msgmerge use fread. and poparser_feed_line() is renamed as
poparser_feed().